+++ sdk/identity/Azure.Identity/src/AzureIdentityEventSource.cs
@@ -31,6 +31,10 @@ namespace Azure.Identity
         private const int ProcessRunnerErrorEvent = 14;
         private const int ProcessRunnerInfoEvent = 15;
         private const int UsernamePasswordCredentialAcquireTokenSilentFailedEvent = 16;
+        private const int TenantIdDiscoveredAndNotUsedEvent = 17;
+        private const int TenantIdDiscoveredAndUsedEvent = 18;
+        internal const string TenantIdDiscoveredAndNotUsedEventMessage = "The service provided an alternate tenantId from what was configured on the credential, but the configured value was used authorization. Configured TenantId: {0}, Alternate TenantId {1}";
+        internal const string TenantIdDiscoveredAndUsedEventMessage = "The service provided an alternate tenantId from what was configured on the credential, and the alternate value was used for authorization. Configured TenantId: {0}, Alternate TenantId {1}";
 
         private AzureIdentityEventSource() : base(EventSourceName) { }
 
@@ -275,6 +279,24 @@ namespace Azure.Identity
         public void UsernamePasswordCredentialAcquireTokenSilentFailed(string error)
         {
             WriteEvent(UsernamePasswordCredentialAcquireTokenSilentFailedEvent, error);
+        }
+
+        [Event(TenantIdDiscoveredAndNotUsedEvent, Level = EventLevel.Informational, Message = TenantIdDiscoveredAndNotUsedEventMessage)]
+        public void TenantIdDiscoveredAndNotUsed(string explicitTenantId, string contextTenantId)
+        {
+            if (IsEnabled(EventLevel.Informational, EventKeywords.All))
+            {
+                WriteEvent(TenantIdDiscoveredAndNotUsedEvent, explicitTenantId, contextTenantId);
+            }
+        }
+
+        [Event(TenantIdDiscoveredAndUsedEvent, Level = EventLevel.Informational, Message = TenantIdDiscoveredAndUsedEventMessage)]
+        public void TenantIdDiscoveredAndUsed(string explicitTenantId, string contextTenantId)
+        {
+            if (IsEnabled(EventLevel.Informational, EventKeywords.All))
+            {
+                WriteEvent(TenantIdDiscoveredAndUsedEvent, explicitTenantId, contextTenantId);
+            }
         }
     }
 }